test(act-rules): revert rule tests back to just being rule.spec.mjs (no allrules.spec.mjs)
Migration to TS and a few tweaks in the next commits.
